assets data 0-schema.sql 1-populate.sql src main java application controller CategoriaController.java OpcaoController.java QuestaoController.java model Categoria.java Opcao.java Questao.java repository CategoriaRepository.java OpcaoRepository.java QuestaoRepository.java Main.java resources application.properties tests generos.http livros.http build.gradle docker-compose.yml settings.gradle

+++ src/main/java/application/repository/CategoriaRepository.java
@@ -1,9 +1,7 @@
 package application.repository;
 
+import application.model.Categoria;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import application.model.Categoria;
-
-public interface CategoriaRepository extends JpaRepository<Categoria, Long> {
-    
+public interface CategoriaRepository extends JpaRepository<Categoria, Integer> {
 }
